_source_extract/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
